Re-show previous frame if nothing has changed
We can detect if nothing has changed from the previous frame by
computing a cheap hash of the draw command stream and if it matches we
reuse the previous contents of the render target we prepare anyway for
resolution scaling.

+++ engine/render/r_texture.cpp
@@ -109,7 +109,7 @@ t_manager_c::~t_manager_c()
 int t_manager_c::GetAsyncCount()
 {
 	std::lock_guard<std::mutex> lock ( mutex );
-	return textureQueue.size();
+	return (int)textureQueue.size();
 }
 
 bool t_manager_c::GetImageInfo(const char* fileName, imageInfo_s* info)
